Add support for testing models on CPU, CUDA, Metal, and Accelerate
This change combines a several changes to enable tests on non-CPU devices:

* Factor out the vector tests to reusable check functions.
* The check functions try all available device types. Ideally this would
  be done using parametrized tests, but this is not supported by rstest
  due to Rust limitations.
* A bunch of changes small changes to fix Metal support (mostly switching
  to supported data types).
* Add `accelerate`, `cuda`, `cudnn`, and `metal` features.

+++ oxidized-transformers/src/layers/embeddings/rotary_embeddings.rs
@@ -216,7 +216,7 @@ impl RotaryEmbeddings {
                 let positions_flat = positions.flatten_all().context(SliceCacheSnafu)?;
                 let max_len = positions_flat
                     .max(0)
-                    .and_then(|xs| xs.to_scalar::<i64>())
+                    .and_then(|xs| xs.to_scalar::<u32>())
                     .context(SliceCacheSnafu)? as usize
                     + 1;
